Begin adding code to add user to database
Created new namsepace, majrix.db.  For now it lives in the same folder
as the main handler.  We'll probably want to change this as we go
forward, but it's ok for now.
handler.clj now has function responsible for calling to add-user in
db.clj and for formatting the response