Start adding tests for handler.clj
compose-respone function has tests and passes
Started tests for the route, but ran into trouble mocking the response
for the call to db.clj. We can use the midje functions 'provided' or
'prerequisite', pay special attention to how you redefine a function in
a different namespace other than the one you're testing.
Tests should start to work once we add the correct parameters to the
mock route call.